pyproject.toml
[project]
name = "severstal-yp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "faiss-cpu==1.12.0",
    "matplotlib==3.3.4",
    "numpy==1.26.4",
    "openpyxl>=3.1.5",
    "pandas==1.2.4",
    "phik==0.12.4",
    "scikit-learn==1.6.1",
    "scipy==1.10.1",
    "seaborn==0.11.1",
    "streamlit>=1.24.1",
    "xlrd>=2.0.2",
]
